# openapi.yaml
openapi: 3.0.0
info:
  title: Second Brain Server API
  description: |
    Comprehensive personal intelligence platform API for knowledge management, creativity, learning, and personal growth.

    ## Features
    - **Stateless Authentication**: JWT-based authentication with no server-side sessions
    - **OAuth Integration**: Google OAuth support with CSRF protection
    - **Database Management**: Create and manage custom databases with flexible schemas
    - **Record Management**: CRUD operations for database records with rich property types
    - **User Management**: Complete user lifecycle management with role-based access control
    - **Rate Limiting**: Built-in rate limiting for security and performance
    - **File Operations**: Import/export data in multiple formats (JSON, CSV, Excel)
    - **Real-time Collaboration**: Share databases with granular permissions

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Login: 5 attempts per 15 minutes
    - Registration: 3 attempts per hour
    - Password Reset: 3 attempts per hour
    - Token Refresh: 10 attempts per 15 minutes
    - OAuth: 20 attempts per 15 minutes
    - General API: 100 requests per 15 minutes

    ## Database Property Types
    - **TEXT**: Single-line text
    - **NUMBER**: Numeric values
...
    - **DATE**: Date and time values
...
  version: 1.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            message:
              type: string
              example: Error description
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Request body validation failed
            statusCode:
              type: integer
              example: 400
            status:
              type: string
              example: error
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: email
                  message:
                    type: string
                    example: Please provide a valid email address
                  value:
                    type: string
                    example: invalid-email
...
paths:
...
  /files:
    get:
      tags: [File Management]
      summary: Get user's files
      description: |
        Retrieve user's uploaded files with filtering and pagination.

        **Features**:
        - Category-based filtering
        - Search by filename or description
        - MIME type filtering
        - Sorting and pagination
      operationId: getUserFiles
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by file category
          schema:
            type: string
        - name: search
          in: query
          description: Search in filename or description
          schema:
            type: string
        - name: mimeType
          in: query
          description: Filter by MIME type
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [name, size, createdAt, updatedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          files:
                            type: array
                            items:
                              $ref: '#/components/schemas/FileResponse'
                          pagination:
                            $ref: '#/components/schemas/PaginationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /files/upload:
    post:
      tags: [File Management]
      summary: Upload a file
      description: |
        Upload a single file to the server.

        **Supported File Types**:
        - Images: JPEG, PNG, GIF, WebP
        - Documents: PDF, Word, Excel, PowerPoint
        - Text: Plain text, CSV, JSON
        - Archives: ZIP

        **File Size Limit**: 50MB
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                description:
                  type: string
                  description: File description
                category:
                  type: string
                  description: File category
                isPublic:
                  type: string
                  enum: ['true', 'false']
                  description: Whether file is publicly accessible
              required:
                - file
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FileResponse'
        '400':
          description: Invalid file or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /files/bulk-upload:
    post:
      tags: [File Management]
      summary: Upload multiple files
      description: |
        Upload multiple files at once (max 10 files).

        **Features**:
        - Batch file processing
        - Consistent metadata for all files
        - Atomic operation (all succeed or all fail)
      operationId: bulkUploadFiles
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
                  description: Files to upload (max 10)
                category:
                  type: string
                  description: Category for all files
                isPublic:
                  type: string
                  enum: ['true', 'false']
                  description: Whether files are publicly accessible
              required:
                - files
      responses:
        '201':
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/FileResponse'
        '400':
          description: Invalid files or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /files/{id}:
    get:
      tags: [File Management]
      summary: Get file details
      description: |
        Retrieve detailed information about a specific file.

        **Features**:
        - File metadata and properties
        - Access control validation
        - Public file access support
      operationId: getFileById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
      responses:
        '200':
          description: File details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags: [File Management]
      summary: Delete file
      description: |
        Delete a file permanently from the system.

        **Features**:
        - Permanent file deletion
        - Storage cleanup
        - Access control validation
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /files/{id}/download:
    get:
      tags: [File Management]
      summary: Download file
      description: |
        Download a file from the server.

        **Features**:
        - Direct file download
        - Proper content headers
        - Access control validation
        - Public file access support
      operationId: downloadFile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Search & Analytics Endpoints
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Get user notifications
      description: |
        Retrieve user's notifications with filtering and pagination.

        **Features**:
        - Read/unread filtering
        - Type-based filtering
        - Pagination support
        - Sorting options
      operationId: getUserNotifications
      security:
        - bearerAuth: []
      parameters:
        - name: isRead
          in: query
          description: Filter by read status
          schema:
            type: boolean
        - name: type
          in: query
          description: Filter by notification type
          schema:
            type: string
            enum: [info, warning, error, success]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, updatedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          notifications:
                            type: array
                            items:
                              $ref: '#/components/schemas/Notification'
                          pagination:
                            $ref: '#/components/schemas/PaginationResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      description: |
        Mark a specific notification as read.

        **Features**:
        - Individual notification marking
        - Read status tracking
        - Timestamp recording
      operationId: markNotificationAsRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/read-all:
    put:
      tags: [Notifications]
      summary: Mark all notifications as read
      description: |
        Mark all user's notifications as read.

        **Features**:
        - Bulk read marking
        - Efficient batch operation
        - Complete notification clearing
      operationId: markAllNotificationsAsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /notifications/{id}:
    delete:
      tags: [Notifications]
      summary: Delete notification
      description: |
        Delete a specific notification permanently.

        **Features**:
        - Permanent notification deletion
        - Individual notification removal
        - Cleanup and optimization
      operationId: deleteNotification
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Notification ID
          schema:
            type: string
      responses:
        '200':
          description: Notification deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Enhanced Database Management Endpoints
...