# src/modules/notifications/controllers/notifications.controller.ts
import { Request, Response, NextFunction } from 'express';
import { catchAsync, sendSuccessResponse, createNotFoundError } from '../../../utils';
import { AuthenticatedRequest } from '../../../middlewares/auth';

/**
 * Get user notifications
 */
export const getUserNotifications = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { 
      isRead, 
      type, 
      page = 1, 
      limit = 20,
      sortBy = 'createdAt',
      sortOrder = 'desc'
    } = req.query;
    
    // TODO: Implement get notifications logic
    const result = {
      notifications: [],
      pagination: {
        total: 0,
        totalPages: 0,
        currentPage: page,
        hasNextPage: false,
        hasPrevPage: false,
        limit
      }
    };

    sendSuccessResponse(res, result, 'Notifications retrieved successfully');
  }
);

/**
 * Mark notification as read
 */
export const markNotificationAsRead = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { id } = req.params;
    
    // TODO: Implement mark as read logic

    sendSuccessResponse(res, null, 'Notification marked as read');
  }
);

/**
 * Mark all notifications as read
 */
export const markAllNotificationsAsRead = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));
    
    // TODO: Implement mark all as read logic

    sendSuccessResponse(res, null, 'All notifications marked as read');
  }
);

/**
 * Delete notification
 */
export const deleteNotification = catchAsync(
  async (req: Request, res: Response, next: NextFunction): Promise<void> => {
    const userId = (req as AuthenticatedRequest).user.userId;
    if (!userId) return next(createNotFoundError('User authentication required'));

    const { id } = req.params;
    
    // TODO: Implement delete notification logic

    sendSuccessResponse(res, null, 'Notification deleted successfully');
  }
);